backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# LangChain dependencies
langchain==0.1.0
langchain-community==0.0.20
langchain-text-splitters==0.0.1
langchain-anthropic==0.1.0
langchain-openai==0.1.0

# Supabase for vector storage
supabase==2.3.0

# Document loaders
pypdf==3.17.0  # For PDF files
